Front end development (#5)
* Adding Bootstrap to the project

* Created Vending Machine Controller and created an index html file.

* front end interface complete

* Adding Bootstrap to the project

* Created Vending Machine Controller and created an index html file.

* front end interface complete

+++ db/migrate/20170112155716_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[5.0]
   def change
     create_table :products do |t|
       t.string :name
-      t.decimal :unit_price
+      t.integer :unit_price
       t.timestamps
     end
   end
